backend/pyproject.toml
[project]
name = "browsecraft-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.1",
  "uvicorn[standard]>=0.35.0",
  "httpx>=0.28.1",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.11.0",
  "browser-use-sdk>=0.10.0",
  "nbtlib>=2.0.4",
  "lmnr>=0.7.44",
  "google-genai>=1.65.0",
  "anthropic>=0.84.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.2",
  "pytest-asyncio>=1.2.0",
  "pytest-cov>=7.0.0",
  "anyio>=4.11.0",
]

[project.scripts]
browsecraft-backend = "browsecraft_backend.main:run"

[build-system]
requires = ["setuptools>=80.9.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
